+++ packages/graphql/tests/tck/experimental/interface-top-level.test.ts
@@ -19,9 +19,9 @@
 
 import { gql } from "graphql-tag";
 import type { DocumentNode } from "graphql";
-import { Neo4jGraphQL } from "../../src";
-import { formatCypher, translateQuery, formatParams } from "./utils/tck-test-utils";
-import { createBearerToken } from "../utils/create-bearer-token";
+import { Neo4jGraphQL } from "../../../src";
+import { formatCypher, translateQuery, formatParams } from "../utils/tck-test-utils";
+import { createBearerToken } from "../../utils/create-bearer-token";
 
 describe("Interface top level operations", () => {
     const secret = "secret";
