+++ packages/graphql/tests/tck/experimental/union-top-level.test.ts
@@ -19,9 +19,9 @@
 
 import { gql } from "graphql-tag";
 import type { DocumentNode } from "graphql";
-import { Neo4jGraphQL } from "../../src";
-import { formatCypher, translateQuery, formatParams } from "./utils/tck-test-utils";
-import { createBearerToken } from "../utils/create-bearer-token";
+import { Neo4jGraphQL } from "../../../src";
+import { formatCypher, translateQuery, formatParams } from "../utils/tck-test-utils";
+import { createBearerToken } from "../../utils/create-bearer-token";
 
 describe("Union top level operations", () => {
     const secret = "secret";
